+++ .changeset/web-extract-composer-input-history.md
@@ -1,0 +1,5 @@
+---
+"@moonshot-ai/kimi-code": patch
+---
+
+Extract the composer's shell-style input-history recall into a reusable composable.
